+++ src/App.tsx
@@ -24,10 +24,6 @@ import '@ionic/react/css/display.css';
 
 /* Theme variables */
 import './theme/variables.css';
-import ClientComponent from "./components/clients/client.component";
-import ClientListComponent from "./components/clients/client-list.component";
-import addClientComponent from "./components/clients/add-client.component";
-import LawyercaseListComponent from "./components/lawyercase/lawyercase-list.component";
 
 
 setupIonicReact();
@@ -39,9 +35,6 @@ const App: React.FC = () => (
         <Layout />
         <IonRouterOutlet class='ion-page' id='main'>
           <Route exact path="/home" component={Home} />
-          <Route exact path="/apiclitest" component={ClientListComponent} />
-          <Route exact path="/apilctest" component={LawyercaseListComponent} />
-          <Route exact path="/add" component={addClientComponent} />
           <Route exact path="/clients" component={Clients} />
           <Route exact path="/records" component={Records} />
           <Route exact path="/" render={() => <Redirect to="/home" />} />
