+++ src/components/Client/AddClient.tsx
@@ -2,6 +2,9 @@ import { IonButton, IonButtons, IonContent, IonHeader, IonIcon, IonInput, IonIte
 import './AddClient.css'
 import { useState } from 'react';
 import { closeOutline, closeSharp } from 'ionicons/icons';
+import IClientData from '../../types/client.type';
+import ClientDataService from "../../services/client.service";
+
 
 interface ModalProps{
     isOpen: boolean;
@@ -10,6 +13,34 @@ interface ModalProps{
 
 const AddClient = (props: ModalProps) => {
     const { isOpen, setIsOpen } = props;
+    const [submitted, setSubmitted] = useState<boolean>(false);
+    const [states, setStates] = useState<IClientData>({
+        name:"",
+        firstname:"",
+        address:"",
+        birthdate:""
+    });
+
+    const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+        setStates({...states, [e.target.name] : e.target.value.trim()})
+    }
+
+    const saveClient = () => {
+        const client: IClientData = {
+            name: states.name,
+            firstname: states.firstname,
+            address: states.address,
+            birthdate: states.birthdate
+        }
+
+        ClientDataService.create(client)
+            .then((res: any) => {
+                console.log(res);
+            })
+            .catch((e: Error) => {
+            console.log(e)
+        })
+    }
 
     return(
         <IonModal isOpen={isOpen} onDidDismiss={() => setIsOpen(false)}>
@@ -25,21 +56,21 @@ const AddClient = (props: ModalProps) => {
                 <form className='ion-padding'>
                     <IonItem>
                         <IonLabel position="floating">Nom</IonLabel>
-                        <IonInput type='text' id='name' required name='nom' ></IonInput>
+                        <IonInput type='text' id='name' required name='nom' onChange={() => handleChange} ></IonInput>
                     </IonItem>
                     <IonItem>
                         <IonLabel position="floating">Prénom</IonLabel>
-                        <IonInput type='text' id='firstname' required name='firstname' ></IonInput>
+                        <IonInput type='text' id='firstname' required name='firstname' onChange={() => handleChange}></IonInput>
                     </IonItem>
                     <IonItem>
                         <IonLabel position="floating">Adresse</IonLabel>
-                        <IonInput type='text' id='name' required name='adresse'></IonInput>
+                        <IonInput type='text' id='name' required name='adresse' onChange={() => handleChange}></IonInput>
                     </IonItem>
                     <IonItem>
                         <IonLabel>Date De Naissance</IonLabel>
-                        <IonInput type='date' id='date' required name='date'></IonInput>
+                        <IonInput type='date' id='date' required name='date' onChange={() => handleChange}></IonInput>
                     </IonItem>
-                    <IonButton expand='block' type='submit'>
+                    <IonButton expand='block' type='submit' onChange={() => saveClient}>
                         Ajouter
                     </IonButton>
                 </form>
