--- src/components/clients/client.component.tsx
@@ -1,10 +1,0 @@
-
-
-const ClientComponent = () => {
-
-    return(
-        <h1>HELLO CLIENT COMPONENT</h1>
-    )
-}
-
-export default ClientComponent;
