--- src/components/lawyercase/add-lawyercase.component.tsx
@@ -1,11 +1,0 @@
-import React from "react"
-
-
-const AddLawyercaseComponent = () =>{
-    return(
-        <>Coucou component
-        </>
-    )
-}
-
-export default AddLawyercaseComponent;
