--- src/components/lawyercase/lawyercase.component.tsx
@@ -1,10 +1,0 @@
-import React from "react"
-
-const LawyercaseComponent = () => {
-    return(
-        <> Coucouc omposant
-        </>
-    )
-}
-
-export default LawyercaseComponent;
