+++ src/pages/Clients/Clients.tsx
@@ -14,29 +14,52 @@ import {
     IonTitle, 
     IonToolbar } from '@ionic/react';
 import { pencilOutline, pencilSharp, trashBinOutline, trashBinSharp, addOutline } from 'ionicons/icons';
-import React, { useState } from 'react';
-import AddClient from './AddClient';
+import React, { useEffect, useState } from 'react';
 import './Clients.css';
-
-interface Client {
-    id: number;
-
-}
+import ClientDataService from "../../services/client.service"
+import IClientData from "../../types/client.type";
+import AddClient from '../../components/Client/AddClient';
 
 
 
 const Clients: React.FC = () => 
 {
-    const [clients, setClients] = useState<Client[]>([]);
     const [isOpen, setIsOpen] = useState(false);
 
     const handleDeleteClient = (id: number) => {
         
     }
 
-    const handleModifyClient = (id: number) => {
-        
+    const handleModifyClient = (id: string) => {
+        deleteClient(id);
     }
+
+    const [clients, setClients] = useState<IClientData[]>([]);
+
+    const retrieveClients = () => {
+        ClientDataService.getAll()
+            .then((response: any) => {
+                setClients(response.data)
+                console.log(response.data);
+            })
+            .catch((e: Error) => {
+                console.log(e);
+            });
+    }
+    const deleteClient = (id:string) => {
+        ClientDataService.delete(id)
+            .then((res:any) => {
+                console.log(res + "A bien été supprimé de la BDD");
+        })
+            .catch((e:Error) => {
+                console.log(e)
+            })
+    }
+
+    useEffect(() => {
+        retrieveClients();  
+
+    }, []);
 
 
     return (
@@ -72,23 +95,26 @@ const Clients: React.FC = () =>
                         <IonCol>Clients</IonCol>
                         <IonCol>Actions</IonCol>
                     </IonRow>
-                    {/* DATA MAP */}
-                    <IonRow>
-                        <IonCol>12/333</IonCol>
-                        <IonCol>En Cours</IonCol>
-                        <IonCol>Abdoul Alim</IonCol>
-                        <IonCol>
-                            <IonButtons>
-                                <IonButton onClick={() => {handleDeleteClient(1)}}>
-                                    <IonIcon ios={trashBinOutline} md={trashBinSharp}></IonIcon>
-                                </IonButton>
-                                <IonButton onClick={() => {handleModifyClient(1)}}>
-                                    <IonIcon ios={pencilOutline} md={pencilSharp}></IonIcon>
-                                </IonButton>
-                            </IonButtons>
-                        </IonCol>
-                    </IonRow>
-                    {/* END */}
+                    {clients.map((client: IClientData, index: number) => {
+
+                        return(
+                        <IonRow>
+                            <IonCol>12/333</IonCol>
+                            <IonCol>En Cours</IonCol>
+                            <IonCol>{client.name + ' '+ client.firstname}</IonCol>
+                            <IonCol>
+                                <IonButtons>
+                                    <IonButton onClick={() => {handleDeleteClient(client.id)}}>
+                                        <IonIcon ios={trashBinOutline} md={trashBinSharp}></IonIcon>
+                                    </IonButton>
+                                    <IonButton onClick={() => {handleModifyClient(client.id)}}>
+                                        <IonIcon ios={pencilOutline} md={pencilSharp}></IonIcon>
+                                    </IonButton>
+                                </IonButtons>
+                            </IonCol>
+                        </IonRow>
+                        )
+                    })}
                 </IonGrid>
             </IonContent>
             <AddClient isOpen={isOpen} setIsOpen={() => setIsOpen(false)}/>
